+++ src/filters/RemoveColor.ts
@@ -3,23 +3,10 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
-
+import { fragmentShader } from './removeColor.shaders';
 export const removeColorDefaultValues: Partial<TClassProperties<RemoveColor>> =
   {
     color: '#FFFFFF',
-    fragmentSource: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      uniform vec4 uLow;
-      uniform vec4 uHigh;
-      varying vec2 vTexCoord;
-      void main() {
-        gl_FragColor = texture2D(uTexture, vTexCoord);
-        if(all(greaterThan(gl_FragColor.rgb,uLow.rgb)) && all(greaterThan(uHigh.rgb,gl_FragColor.rgb))) {
-          gl_FragColor.a = 0.0;
-        }
-      }
-      `,
     distance: 0.02,
     useAlpha: false,
   };
@@ -55,6 +42,10 @@ export class RemoveColor extends BaseFilter {
   declare useAlpha: boolean;
 
   static defaults = removeColorDefaultValues;
+
+  getFragmentShader() {
+    return fragmentShader;
+  }
 
   /**
    * Applies filter to canvas element
