+++ src/filters/removeColor.shaders.ts
@@ -1,0 +1,13 @@
+export const fragmentShader = `
+precision highp float;
+uniform sampler2D uTexture;
+uniform vec4 uLow;
+uniform vec4 uHigh;
+varying vec2 vTexCoord;
+void main() {
+  gl_FragColor = texture2D(uTexture, vTexCoord);
+  if(all(greaterThan(gl_FragColor.rgb,uLow.rgb)) && all(greaterThan(uHigh.rgb,gl_FragColor.rgb))) {
+    gl_FragColor.a = 0.0;
+  }
+}
+`;
